fix(proof): share full Unicode case semantics across backends
