+++ src/codegen/lean/builtins.rs
@@ -194,8 +194,8 @@ pub fn emit_builtin_call(
         StringSplit => format!("AverString.split {} {}", p(&a[0]), p(&a[1])),
         StringJoin => format!("String.intercalate {} {}", p(&a[1]), p(&a[0])),
         StringReplace => format!("{}.replace {} {}", p(&a[0]), p(&a[1]), p(&a[2])),
-        StringToUpper => format!("{}.toUpper", p(&a[0])),
-        StringToLower => format!("{}.toLower", p(&a[0])),
+        StringToUpper => format!("AverUnicodeCase.toUpper {}", p(&a[0])),
+        StringToLower => format!("AverUnicodeCase.toLower {}", p(&a[0])),
         StringFromInt => format!("String.fromInt {}", p(&a[0])),
         StringFromFloat => format!("String.fromFloat {}", p(&a[0])),
         StringFromBool => format!("String.fromBool {}", p(&a[0])),
