+++ src/codegen/wasm_gc/builtins/case_tables.rs
@@ -1,2158 +1,12 @@
-//! Unicode case-mapping tables for the wasm-gc `String.toLower` /
-//! `String.toUpper` helpers, plus the passive-data-segment blob the
-//! helpers read them from.
-//!
-//! # Where the numbers come from
-//!
-//! Every table below is DERIVED FROM RUST'S OWN `char::to_lowercase`
-//! / `char::to_uppercase` by probing all 1 112 064 Unicode scalars —
-//! no new dependency, and the wasm-gc backend maps exactly what the VM
-//! (`src/types/string.rs`) and the generated Rust
-//! (`src/codegen/rust/from_mir.rs`) map. The two context predicates the
-//! final-sigma rule needs are probed the same way, through
-//! `str::to_lowercase` itself:
-//!
-//! ```text
-//! cased(c)      := format!("{c}Σ").to_lowercase().ends_with('ς')
-//! ignorable(c)  := !cased(c) && format!("a{c}Σ").to_lowercase().ends_with('ς')
-//! ```
-//!
-//! These deliberately return the EFFECTIVE predicates in the order
-//! `str::to_lowercase` applies them (`skip_while(Case_Ignorable).next()
-//! .map(Cased)`): a scalar that Unicode marks both Cased and
-//! Case_Ignorable — modifier letters such as U+02B0 — comes out of the
-//! probes as ignorable-only, which is what the helper must implement.
-//! The two sets are therefore disjoint by construction.
-//!
-//! # Regenerating
-//!
-//! The constants are committed, and `case_tables_match_std` re-derives
-//! them from std on every test run — so a toolchain whose Unicode
-//! version moved turns that test red instead of silently shipping stale
-//! mappings. To refresh:
-//!
-//! ```text
-//! cargo test -p aver-lang --lib --features wasm,wasip2 \
-//!     dump_case_tables -- --ignored --nocapture
-//! ```
-//!
-//! and paste the printed constants over the ones below.
+//! Pack the shared Unicode case tables into a wasm passive data segment.
 
-/// One run of scalars that all shift by the same amount:
-/// `(first, last, delta)`, sorted by `first`, disjoint.
-pub(in crate::codegen::wasm_gc) type SimpleRun = (u32, u32, i32);
-
-/// One scalar whose case mapping is longer than a single scalar:
-/// `(from, [to0, to1, to2])`, with `to2 == 0` meaning "two scalars".
-pub(in crate::codegen::wasm_gc) type Expansion = (u32, [u32; 3]);
-
-/// An inclusive scalar range, sorted and disjoint.
-pub(in crate::codegen::wasm_gc) type Range = (u32, u32);
-
-/// Largest ratio of output UTF-8 bytes to input UTF-8 bytes over all
-/// scalars, rounded up — the over-allocation factor `String.toLower`
-/// uses for its output buffer.
-pub(in crate::codegen::wasm_gc) const MAX_LOWER_GROWTH: u32 = 2;
-
-/// Same for `String.toUpper`. U+0390 and U+03B0 are the worst case:
-/// one 2-byte scalar becomes three 2-byte scalars.
-pub(in crate::codegen::wasm_gc) const MAX_UPPER_GROWTH: u32 = 3;
-
-// @generated by `dump_case_tables` — see the module docs.
-pub(in crate::codegen::wasm_gc) const LOWER_SIMPLE: &[SimpleRun] = &[
-    (65, 90, 32),
-    (192, 214, 32),
-    (216, 222, 32),
-    (256, 256, 1),
-    (258, 258, 1),
-    (260, 260, 1),
-    (262, 262, 1),
-    (264, 264, 1),
-    (266, 266, 1),
-    (268, 268, 1),
-    (270, 270, 1),
-    (272, 272, 1),
-    (274, 274, 1),
-    (276, 276, 1),
-    (278, 278, 1),
-    (280, 280, 1),
-    (282, 282, 1),
-    (284, 284, 1),
-    (286, 286, 1),
-    (288, 288, 1),
-    (290, 290, 1),
-    (292, 292, 1),
-    (294, 294, 1),
-    (296, 296, 1),
-    (298, 298, 1),
-    (300, 300, 1),
-    (302, 302, 1),
-    (306, 306, 1),
-    (308, 308, 1),
-    (310, 310, 1),
-    (313, 313, 1),
-    (315, 315, 1),
-    (317, 317, 1),
-    (319, 319, 1),
-    (321, 321, 1),
-    (323, 323, 1),
-    (325, 325, 1),
-    (327, 327, 1),
-    (330, 330, 1),
-    (332, 332, 1),
-    (334, 334, 1),
-    (336, 336, 1),
-    (338, 338, 1),
-    (340, 340, 1),
-    (342, 342, 1),
-    (344, 344, 1),
-    (346, 346, 1),
-    (348, 348, 1),
-    (350, 350, 1),
-    (352, 352, 1),
-    (354, 354, 1),
-    (356, 356, 1),
-    (358, 358, 1),
-    (360, 360, 1),
-    (362, 362, 1),
-    (364, 364, 1),
-    (366, 366, 1),
-    (368, 368, 1),
-    (370, 370, 1),
-    (372, 372, 1),
-    (374, 374, 1),
-    (376, 376, -121),
-    (377, 377, 1),
-    (379, 379, 1),
-    (381, 381, 1),
-    (385, 385, 210),
-    (386, 386, 1),
-    (388, 388, 1),
-    (390, 390, 206),
-    (391, 391, 1),
-    (393, 394, 205),
-    (395, 395, 1),
-    (398, 398, 79),
-    (399, 399, 202),
-    (400, 400, 203),
-    (401, 401, 1),
-    (403, 403, 205),
-    (404, 404, 207),
-    (406, 406, 211),
-    (407, 407, 209),
-    (408, 408, 1),
-    (412, 412, 211),
-    (413, 413, 213),
-    (415, 415, 214),
-    (416, 416, 1),
-    (418, 418, 1),
-    (420, 420, 1),
-    (422, 422, 218),
-    (423, 423, 1),
-    (425, 425, 218),
-    (428, 428, 1),
-    (430, 430, 218),
-    (431, 431, 1),
-    (433, 434, 217),
-    (435, 435, 1),
-    (437, 437, 1),
-    (439, 439, 219),
-    (440, 440, 1),
-    (444, 444, 1),
-    (452, 452, 2),
-    (453, 453, 1),
-    (455, 455, 2),
-    (456, 456, 1),
-    (458, 458, 2),
-    (459, 459, 1),
-    (461, 461, 1),
-    (463, 463, 1),
-    (465, 465, 1),
-    (467, 467, 1),
-    (469, 469, 1),
-    (471, 471, 1),
-    (473, 473, 1),
-    (475, 475, 1),
-    (478, 478, 1),
-    (480, 480, 1),
-    (482, 482, 1),
-    (484, 484, 1),
-    (486, 486, 1),
-    (488, 488, 1),
-    (490, 490, 1),
-    (492, 492, 1),
-    (494, 494, 1),
-    (497, 497, 2),
-    (498, 498, 1),
-    (500, 500, 1),
-    (502, 502, -97),
-    (503, 503, -56),
-    (504, 504, 1),
-    (506, 506, 1),
-    (508, 508, 1),
-    (510, 510, 1),
-    (512, 512, 1),
-    (514, 514, 1),
-    (516, 516, 1),
-    (518, 518, 1),
-    (520, 520, 1),
-    (522, 522, 1),
-    (524, 524, 1),
-    (526, 526, 1),
-    (528, 528, 1),
-    (530, 530, 1),
-    (532, 532, 1),
-    (534, 534, 1),
-    (536, 536, 1),
-    (538, 538, 1),
-    (540, 540, 1),
-    (542, 542, 1),
-    (544, 544, -130),
-    (546, 546, 1),
-    (548, 548, 1),
-    (550, 550, 1),
-    (552, 552, 1),
-    (554, 554, 1),
-    (556, 556, 1),
-    (558, 558, 1),
-    (560, 560, 1),
-    (562, 562, 1),
-    (570, 570, 10795),
-    (571, 571, 1),
-    (573, 573, -163),
-    (574, 574, 10792),
-    (577, 577, 1),
-    (579, 579, -195),
-    (580, 580, 69),
-    (581, 581, 71),
-    (582, 582, 1),
-    (584, 584, 1),
-    (586, 586, 1),
-    (588, 588, 1),
-    (590, 590, 1),
-    (880, 880, 1),
-    (882, 882, 1),
-    (886, 886, 1),
-    (895, 895, 116),
-    (902, 902, 38),
-    (904, 906, 37),
-    (908, 908, 64),
-    (910, 911, 63),
-    (913, 929, 32),
-    (931, 939, 32),
-    (975, 975, 8),
-    (984, 984, 1),
-    (986, 986, 1),
-    (988, 988, 1),
-    (990, 990, 1),
-    (992, 992, 1),
-    (994, 994, 1),
-    (996, 996, 1),
-    (998, 998, 1),
-    (1000, 1000, 1),
-    (1002, 1002, 1),
-    (1004, 1004, 1),
-    (1006, 1006, 1),
-    (1012, 1012, -60),
-    (1015, 1015, 1),
-    (1017, 1017, -7),
-    (1018, 1018, 1),
-    (1021, 1023, -130),
-    (1024, 1039, 80),
-    (1040, 1071, 32),
-    (1120, 1120, 1),
-    (1122, 1122, 1),
-    (1124, 1124, 1),
-    (1126, 1126, 1),
-    (1128, 1128, 1),
-    (1130, 1130, 1),
-    (1132, 1132, 1),
-    (1134, 1134, 1),
-    (1136, 1136, 1),
-    (1138, 1138, 1),
-    (1140, 1140, 1),
-    (1142, 1142, 1),
-    (1144, 1144, 1),
-    (1146, 1146, 1),
-    (1148, 1148, 1),
-    (1150, 1150, 1),
-    (1152, 1152, 1),
-    (1162, 1162, 1),
-    (1164, 1164, 1),
-    (1166, 1166, 1),
-    (1168, 1168, 1),
-    (1170, 1170, 1),
-    (1172, 1172, 1),
-    (1174, 1174, 1),
-    (1176, 1176, 1),
-    (1178, 1178, 1),
-    (1180, 1180, 1),
-    (1182, 1182, 1),
-    (1184, 1184, 1),
-    (1186, 1186, 1),
-    (1188, 1188, 1),
-    (1190, 1190, 1),
-    (1192, 1192, 1),
-    (1194, 1194, 1),
-    (1196, 1196, 1),
-    (1198, 1198, 1),
-    (1200, 1200, 1),
-    (1202, 1202, 1),
-    (1204, 1204, 1),
-    (1206, 1206, 1),
-    (1208, 1208, 1),
-    (1210, 1210, 1),
-    (1212, 1212, 1),
-    (1214, 1214, 1),
-    (1216, 1216, 15),
-    (1217, 1217, 1),
-    (1219, 1219, 1),
-    (1221, 1221, 1),
-    (1223, 1223, 1),
-    (1225, 1225, 1),
-    (1227, 1227, 1),
-    (1229, 1229, 1),
-    (1232, 1232, 1),
-    (1234, 1234, 1),
-    (1236, 1236, 1),
-    (1238, 1238, 1),
-    (1240, 1240, 1),
-    (1242, 1242, 1),
-    (1244, 1244, 1),
-    (1246, 1246, 1),
-    (1248, 1248, 1),
-    (1250, 1250, 1),
-    (1252, 1252, 1),
-    (1254, 1254, 1),
-    (1256, 1256, 1),
-    (1258, 1258, 1),
-    (1260, 1260, 1),
-    (1262, 1262, 1),
-    (1264, 1264, 1),
-    (1266, 1266, 1),
-    (1268, 1268, 1),
-    (1270, 1270, 1),
-    (1272, 1272, 1),
-    (1274, 1274, 1),
-    (1276, 1276, 1),
-    (1278, 1278, 1),
-    (1280, 1280, 1),
-    (1282, 1282, 1),
-    (1284, 1284, 1),
-    (1286, 1286, 1),
-    (1288, 1288, 1),
-    (1290, 1290, 1),
-    (1292, 1292, 1),
-    (1294, 1294, 1),
-    (1296, 1296, 1),
-    (1298, 1298, 1),
-    (1300, 1300, 1),
-    (1302, 1302, 1),
-    (1304, 1304, 1),
-    (1306, 1306, 1),
-    (1308, 1308, 1),
-    (1310, 1310, 1),
-    (1312, 1312, 1),
-    (1314, 1314, 1),
-    (1316, 1316, 1),
-    (1318, 1318, 1),
-    (1320, 1320, 1),
-    (1322, 1322, 1),
-    (1324, 1324, 1),
-    (1326, 1326, 1),
-    (1329, 1366, 48),
-    (4256, 4293, 7264),
-    (4295, 4295, 7264),
-    (4301, 4301, 7264),
-    (5024, 5103, 38864),
-    (5104, 5109, 8),
-    (7305, 7305, 1),
-    (7312, 7354, -3008),
-    (7357, 7359, -3008),
-    (7680, 7680, 1),
-    (7682, 7682, 1),
-    (7684, 7684, 1),
-    (7686, 7686, 1),
-    (7688, 7688, 1),
-    (7690, 7690, 1),
-    (7692, 7692, 1),
-    (7694, 7694, 1),
-    (7696, 7696, 1),
-    (7698, 7698, 1),
-    (7700, 7700, 1),
-    (7702, 7702, 1),
-    (7704, 7704, 1),
-    (7706, 7706, 1),
-    (7708, 7708, 1),
-    (7710, 7710, 1),
-    (7712, 7712, 1),
-    (7714, 7714, 1),
-    (7716, 7716, 1),
-    (7718, 7718, 1),
-    (7720, 7720, 1),
-    (7722, 7722, 1),
-    (7724, 7724, 1),
-    (7726, 7726, 1),
-    (7728, 7728, 1),
-    (7730, 7730, 1),
-    (7732, 7732, 1),
-    (7734, 7734, 1),
-    (7736, 7736, 1),
-    (7738, 7738, 1),
-    (7740, 7740, 1),
-    (7742, 7742, 1),
-    (7744, 7744, 1),
-    (7746, 7746, 1),
-    (7748, 7748, 1),
-    (7750, 7750, 1),
-    (7752, 7752, 1),
-    (7754, 7754, 1),
-    (7756, 7756, 1),
-    (7758, 7758, 1),
-    (7760, 7760, 1),
-    (7762, 7762, 1),
-    (7764, 7764, 1),
-    (7766, 7766, 1),
-    (7768, 7768, 1),
-    (7770, 7770, 1),
-    (7772, 7772, 1),
-    (7774, 7774, 1),
-    (7776, 7776, 1),
-    (7778, 7778, 1),
-    (7780, 7780, 1),
-    (7782, 7782, 1),
-    (7784, 7784, 1),
-    (7786, 7786, 1),
-    (7788, 7788, 1),
-    (7790, 7790, 1),
-    (7792, 7792, 1),
-    (7794, 7794, 1),
-    (7796, 7796, 1),
-    (7798, 7798, 1),
-    (7800, 7800, 1),
-    (7802, 7802, 1),
-    (7804, 7804, 1),
-    (7806, 7806, 1),
-    (7808, 7808, 1),
-    (7810, 7810, 1),
-    (7812, 7812, 1),
-    (7814, 7814, 1),
-    (7816, 7816, 1),
-    (7818, 7818, 1),
-    (7820, 7820, 1),
-    (7822, 7822, 1),
-    (7824, 7824, 1),
-    (7826, 7826, 1),
-    (7828, 7828, 1),
-    (7838, 7838, -7615),
-    (7840, 7840, 1),
-    (7842, 7842, 1),
-    (7844, 7844, 1),
-    (7846, 7846, 1),
-    (7848, 7848, 1),
-    (7850, 7850, 1),
-    (7852, 7852, 1),
-    (7854, 7854, 1),
-    (7856, 7856, 1),
-    (7858, 7858, 1),
-    (7860, 7860, 1),
-    (7862, 7862, 1),
-    (7864, 7864, 1),
-    (7866, 7866, 1),
-    (7868, 7868, 1),
-    (7870, 7870, 1),
-    (7872, 7872, 1),
-    (7874, 7874, 1),
-    (7876, 7876, 1),
-    (7878, 7878, 1),
-    (7880, 7880, 1),
-    (7882, 7882, 1),
-    (7884, 7884, 1),
-    (7886, 7886, 1),
-    (7888, 7888, 1),
-    (7890, 7890, 1),
-    (7892, 7892, 1),
-    (7894, 7894, 1),
-    (7896, 7896, 1),
-    (7898, 7898, 1),
-    (7900, 7900, 1),
-    (7902, 7902, 1),
-    (7904, 7904, 1),
-    (7906, 7906, 1),
-    (7908, 7908, 1),
-    (7910, 7910, 1),
-    (7912, 7912, 1),
-    (7914, 7914, 1),
-    (7916, 7916, 1),
-    (7918, 7918, 1),
-    (7920, 7920, 1),
-    (7922, 7922, 1),
-    (7924, 7924, 1),
-    (7926, 7926, 1),
-    (7928, 7928, 1),
-    (7930, 7930, 1),
-    (7932, 7932, 1),
-    (7934, 7934, 1),
-    (7944, 7951, -8),
-    (7960, 7965, -8),
-    (7976, 7983, -8),
-    (7992, 7999, -8),
-    (8008, 8013, -8),
-    (8025, 8025, -8),
-    (8027, 8027, -8),
-    (8029, 8029, -8),
-    (8031, 8031, -8),
-    (8040, 8047, -8),
-    (8072, 8079, -8),
-    (8088, 8095, -8),
-    (8104, 8111, -8),
-    (8120, 8121, -8),
-    (8122, 8123, -74),
-    (8124, 8124, -9),
-    (8136, 8139, -86),
-    (8140, 8140, -9),
-    (8152, 8153, -8),
-    (8154, 8155, -100),
-    (8168, 8169, -8),
-    (8170, 8171, -112),
-    (8172, 8172, -7),
-    (8184, 8185, -128),
-    (8186, 8187, -126),
-    (8188, 8188, -9),
-    (8486, 8486, -7517),
-    (8490, 8490, -8383),
-    (8491, 8491, -8262),
-    (8498, 8498, 28),
-    (8544, 8559, 16),
-    (8579, 8579, 1),
-    (9398, 9423, 26),
-    (11264, 11311, 48),
-    (11360, 11360, 1),
-    (11362, 11362, -10743),
-    (11363, 11363, -3814),
-    (11364, 11364, -10727),
-    (11367, 11367, 1),
-    (11369, 11369, 1),
-    (11371, 11371, 1),
-    (11373, 11373, -10780),
-    (11374, 11374, -10749),
-    (11375, 11375, -10783),
-    (11376, 11376, -10782),
-    (11378, 11378, 1),
-    (11381, 11381, 1),
-    (11390, 11391, -10815),
-    (11392, 11392, 1),
-    (11394, 11394, 1),
-    (11396, 11396, 1),
-    (11398, 11398, 1),
-    (11400, 11400, 1),
-    (11402, 11402, 1),
-    (11404, 11404, 1),
-    (11406, 11406, 1),
-    (11408, 11408, 1),
-    (11410, 11410, 1),
-    (11412, 11412, 1),
-    (11414, 11414, 1),
-    (11416, 11416, 1),
-    (11418, 11418, 1),
-    (11420, 11420, 1),
-    (11422, 11422, 1),
-    (11424, 11424, 1),
-    (11426, 11426, 1),
-    (11428, 11428, 1),
-    (11430, 11430, 1),
-    (11432, 11432, 1),
-    (11434, 11434, 1),
-    (11436, 11436, 1),
-    (11438, 11438, 1),
-    (11440, 11440, 1),
-    (11442, 11442, 1),
-    (11444, 11444, 1),
-    (11446, 11446, 1),
-    (11448, 11448, 1),
-    (11450, 11450, 1),
-    (11452, 11452, 1),
-    (11454, 11454, 1),
-    (11456, 11456, 1),
-    (11458, 11458, 1),
-    (11460, 11460, 1),
-    (11462, 11462, 1),
-    (11464, 11464, 1),
-    (11466, 11466, 1),
-    (11468, 11468, 1),
-    (11470, 11470, 1),
-    (11472, 11472, 1),
-    (11474, 11474, 1),
-    (11476, 11476, 1),
-    (11478, 11478, 1),
-    (11480, 11480, 1),
-    (11482, 11482, 1),
-    (11484, 11484, 1),
-    (11486, 11486, 1),
-    (11488, 11488, 1),
-    (11490, 11490, 1),
-    (11499, 11499, 1),
-    (11501, 11501, 1),
-    (11506, 11506, 1),
-    (42560, 42560, 1),
-    (42562, 42562, 1),
-    (42564, 42564, 1),
-    (42566, 42566, 1),
-    (42568, 42568, 1),
-    (42570, 42570, 1),
-    (42572, 42572, 1),
-    (42574, 42574, 1),
-    (42576, 42576, 1),
-    (42578, 42578, 1),
-    (42580, 42580, 1),
-    (42582, 42582, 1),
-    (42584, 42584, 1),
-    (42586, 42586, 1),
-    (42588, 42588, 1),
-    (42590, 42590, 1),
-    (42592, 42592, 1),
-    (42594, 42594, 1),
-    (42596, 42596, 1),
-    (42598, 42598, 1),
-    (42600, 42600, 1),
-    (42602, 42602, 1),
-    (42604, 42604, 1),
-    (42624, 42624, 1),
-    (42626, 42626, 1),
-    (42628, 42628, 1),
-    (42630, 42630, 1),
-    (42632, 42632, 1),
-    (42634, 42634, 1),
-    (42636, 42636, 1),
-    (42638, 42638, 1),
-    (42640, 42640, 1),
-    (42642, 42642, 1),
-    (42644, 42644, 1),
-    (42646, 42646, 1),
-    (42648, 42648, 1),
-    (42650, 42650, 1),
-    (42786, 42786, 1),
-    (42788, 42788, 1),
-    (42790, 42790, 1),
-    (42792, 42792, 1),
-    (42794, 42794, 1),
-    (42796, 42796, 1),
-    (42798, 42798, 1),
-    (42802, 42802, 1),
-    (42804, 42804, 1),
-    (42806, 42806, 1),
-    (42808, 42808, 1),
-    (42810, 42810, 1),
-    (42812, 42812, 1),
-    (42814, 42814, 1),
-    (42816, 42816, 1),
-    (42818, 42818, 1),
-    (42820, 42820, 1),
-    (42822, 42822, 1),
-    (42824, 42824, 1),
-    (42826, 42826, 1),
-    (42828, 42828, 1),
-    (42830, 42830, 1),
-    (42832, 42832, 1),
-    (42834, 42834, 1),
-    (42836, 42836, 1),
-    (42838, 42838, 1),
-    (42840, 42840, 1),
-    (42842, 42842, 1),
-    (42844, 42844, 1),
-    (42846, 42846, 1),
-    (42848, 42848, 1),
-    (42850, 42850, 1),
-    (42852, 42852, 1),
-    (42854, 42854, 1),
-    (42856, 42856, 1),
-    (42858, 42858, 1),
-    (42860, 42860, 1),
-    (42862, 42862, 1),
-    (42873, 42873, 1),
-    (42875, 42875, 1),
-    (42877, 42877, -35332),
-    (42878, 42878, 1),
-    (42880, 42880, 1),
-    (42882, 42882, 1),
-    (42884, 42884, 1),
-    (42886, 42886, 1),
-    (42891, 42891, 1),
-    (42893, 42893, -42280),
-    (42896, 42896, 1),
-    (42898, 42898, 1),
-    (42902, 42902, 1),
-    (42904, 42904, 1),
-    (42906, 42906, 1),
-    (42908, 42908, 1),
-    (42910, 42910, 1),
-    (42912, 42912, 1),
-    (42914, 42914, 1),
-    (42916, 42916, 1),
-    (42918, 42918, 1),
-    (42920, 42920, 1),
-    (42922, 42922, -42308),
-    (42923, 42923, -42319),
-    (42924, 42924, -42315),
-    (42925, 42925, -42305),
-    (42926, 42926, -42308),
-    (42928, 42928, -42258),
-    (42929, 42929, -42282),
-    (42930, 42930, -42261),
-    (42931, 42931, 928),
-    (42932, 42932, 1),
-    (42934, 42934, 1),
-    (42936, 42936, 1),
-    (42938, 42938, 1),
-    (42940, 42940, 1),
-    (42942, 42942, 1),
-    (42944, 42944, 1),
-    (42946, 42946, 1),
-    (42948, 42948, -48),
-    (42949, 42949, -42307),
-    (42950, 42950, -35384),
-    (42951, 42951, 1),
-    (42953, 42953, 1),
-    (42955, 42955, -42343),
-    (42956, 42956, 1),
-    (42958, 42958, 1),
-    (42960, 42960, 1),
-    (42962, 42962, 1),
-    (42964, 42964, 1),
-    (42966, 42966, 1),
-    (42968, 42968, 1),
-    (42970, 42970, 1),
-    (42972, 42972, -42561),
-    (42997, 42997, 1),
-    (65313, 65338, 32),
-    (66560, 66599, 40),
-    (66736, 66771, 40),
-    (66928, 66938, 39),
-    (66940, 66954, 39),
-    (66956, 66962, 39),
-    (66964, 66965, 39),
-    (68736, 68786, 64),
-    (68944, 68965, 32),
-    (71840, 71871, 32),
-    (93760, 93791, 32),
-    (93856, 93880, 27),
-    (125184, 125217, 34),
-];
-pub(in crate::codegen::wasm_gc) const UPPER_SIMPLE: &[SimpleRun] = &[
-    (97, 122, -32),
-    (181, 181, 743),
-    (224, 246, -32),
-    (248, 254, -32),
-    (255, 255, 121),
-    (257, 257, -1),
-    (259, 259, -1),
-    (261, 261, -1),
-    (263, 263, -1),
-    (265, 265, -1),
-    (267, 267, -1),
-    (269, 269, -1),
-    (271, 271, -1),
-    (273, 273, -1),
-    (275, 275, -1),
-    (277, 277, -1),
-    (279, 279, -1),
-    (281, 281, -1),
-    (283, 283, -1),
-    (285, 285, -1),
-    (287, 287, -1),
-    (289, 289, -1),
-    (291, 291, -1),
-    (293, 293, -1),
-    (295, 295, -1),
-    (297, 297, -1),
-    (299, 299, -1),
-    (301, 301, -1),
-    (303, 303, -1),
-    (305, 305, -232),
-    (307, 307, -1),
-    (309, 309, -1),
-    (311, 311, -1),
-    (314, 314, -1),
-    (316, 316, -1),
-    (318, 318, -1),
-    (320, 320, -1),
-    (322, 322, -1),
-    (324, 324, -1),
-    (326, 326, -1),
-    (328, 328, -1),
-    (331, 331, -1),
-    (333, 333, -1),
-    (335, 335, -1),
-    (337, 337, -1),
-    (339, 339, -1),
-    (341, 341, -1),
-    (343, 343, -1),
-    (345, 345, -1),
-    (347, 347, -1),
-    (349, 349, -1),
-    (351, 351, -1),
-    (353, 353, -1),
-    (355, 355, -1),
-    (357, 357, -1),
-    (359, 359, -1),
-    (361, 361, -1),
-    (363, 363, -1),
-    (365, 365, -1),
-    (367, 367, -1),
-    (369, 369, -1),
-    (371, 371, -1),
-    (373, 373, -1),
-    (375, 375, -1),
-    (378, 378, -1),
-    (380, 380, -1),
-    (382, 382, -1),
-    (383, 383, -300),
-    (384, 384, 195),
-    (387, 387, -1),
-    (389, 389, -1),
-    (392, 392, -1),
-    (396, 396, -1),
-    (402, 402, -1),
-    (405, 405, 97),
-    (409, 409, -1),
-    (410, 410, 163),
-    (411, 411, 42561),
-    (414, 414, 130),
-    (417, 417, -1),
-    (419, 419, -1),
-    (421, 421, -1),
-    (424, 424, -1),
-    (429, 429, -1),
-    (432, 432, -1),
-    (436, 436, -1),
-    (438, 438, -1),
-    (441, 441, -1),
-    (445, 445, -1),
-    (447, 447, 56),
-    (453, 453, -1),
-    (454, 454, -2),
-    (456, 456, -1),
-    (457, 457, -2),
-    (459, 459, -1),
-    (460, 460, -2),
-    (462, 462, -1),
-    (464, 464, -1),
-    (466, 466, -1),
-    (468, 468, -1),
-    (470, 470, -1),
-    (472, 472, -1),
-    (474, 474, -1),
-    (476, 476, -1),
-    (477, 477, -79),
-    (479, 479, -1),
-    (481, 481, -1),
-    (483, 483, -1),
-    (485, 485, -1),
-    (487, 487, -1),
-    (489, 489, -1),
-    (491, 491, -1),
-    (493, 493, -1),
-    (495, 495, -1),
-    (498, 498, -1),
-    (499, 499, -2),
-    (501, 501, -1),
-    (505, 505, -1),
-    (507, 507, -1),
-    (509, 509, -1),
-    (511, 511, -1),
-    (513, 513, -1),
-    (515, 515, -1),
-    (517, 517, -1),
-    (519, 519, -1),
-    (521, 521, -1),
-    (523, 523, -1),
-    (525, 525, -1),
-    (527, 527, -1),
-    (529, 529, -1),
-    (531, 531, -1),
-    (533, 533, -1),
-    (535, 535, -1),
-    (537, 537, -1),
-    (539, 539, -1),
-    (541, 541, -1),
-    (543, 543, -1),
-    (547, 547, -1),
-    (549, 549, -1),
-    (551, 551, -1),
-    (553, 553, -1),
-    (555, 555, -1),
-    (557, 557, -1),
-    (559, 559, -1),
-    (561, 561, -1),
-    (563, 563, -1),
-    (572, 572, -1),
-    (575, 576, 10815),
-    (578, 578, -1),
-    (583, 583, -1),
-    (585, 585, -1),
-    (587, 587, -1),
-    (589, 589, -1),
-    (591, 591, -1),
-    (592, 592, 10783),
-    (593, 593, 10780),
-    (594, 594, 10782),
-    (595, 595, -210),
-    (596, 596, -206),
-    (598, 599, -205),
-    (601, 601, -202),
-    (603, 603, -203),
-    (604, 604, 42319),
-    (608, 608, -205),
-    (609, 609, 42315),
-    (611, 611, -207),
-    (612, 612, 42343),
-    (613, 613, 42280),
-    (614, 614, 42308),
-    (616, 616, -209),
-    (617, 617, -211),
-    (618, 618, 42308),
-    (619, 619, 10743),
-    (620, 620, 42305),
-    (623, 623, -211),
-    (625, 625, 10749),
-    (626, 626, -213),
-    (629, 629, -214),
-    (637, 637, 10727),
-    (640, 640, -218),
-    (642, 642, 42307),
-    (643, 643, -218),
-    (647, 647, 42282),
-    (648, 648, -218),
-    (649, 649, -69),
-    (650, 651, -217),
-    (652, 652, -71),
-    (658, 658, -219),
-    (669, 669, 42261),
-    (670, 670, 42258),
-    (837, 837, 84),
-    (881, 881, -1),
-    (883, 883, -1),
-    (887, 887, -1),
-    (891, 893, 130),
-    (940, 940, -38),
-    (941, 943, -37),
-    (945, 961, -32),
-    (962, 962, -31),
-    (963, 971, -32),
-    (972, 972, -64),
-    (973, 974, -63),
-    (976, 976, -62),
-    (977, 977, -57),
-    (981, 981, -47),
-    (982, 982, -54),
-    (983, 983, -8),
-    (985, 985, -1),
-    (987, 987, -1),
-    (989, 989, -1),
-    (991, 991, -1),
-    (993, 993, -1),
-    (995, 995, -1),
-    (997, 997, -1),
-    (999, 999, -1),
-    (1001, 1001, -1),
-    (1003, 1003, -1),
-    (1005, 1005, -1),
-    (1007, 1007, -1),
-    (1008, 1008, -86),
-    (1009, 1009, -80),
-    (1010, 1010, 7),
-    (1011, 1011, -116),
-    (1013, 1013, -96),
-    (1016, 1016, -1),
-    (1019, 1019, -1),
-    (1072, 1103, -32),
-    (1104, 1119, -80),
-    (1121, 1121, -1),
-    (1123, 1123, -1),
-    (1125, 1125, -1),
-    (1127, 1127, -1),
-    (1129, 1129, -1),
-    (1131, 1131, -1),
-    (1133, 1133, -1),
-    (1135, 1135, -1),
-    (1137, 1137, -1),
-    (1139, 1139, -1),
-    (1141, 1141, -1),
-    (1143, 1143, -1),
-    (1145, 1145, -1),
-    (1147, 1147, -1),
-    (1149, 1149, -1),
-    (1151, 1151, -1),
-    (1153, 1153, -1),
-    (1163, 1163, -1),
-    (1165, 1165, -1),
-    (1167, 1167, -1),
-    (1169, 1169, -1),
-    (1171, 1171, -1),
-    (1173, 1173, -1),
-    (1175, 1175, -1),
-    (1177, 1177, -1),
-    (1179, 1179, -1),
-    (1181, 1181, -1),
-    (1183, 1183, -1),
-    (1185, 1185, -1),
-    (1187, 1187, -1),
-    (1189, 1189, -1),
-    (1191, 1191, -1),
-    (1193, 1193, -1),
-    (1195, 1195, -1),
-    (1197, 1197, -1),
-    (1199, 1199, -1),
-    (1201, 1201, -1),
-    (1203, 1203, -1),
-    (1205, 1205, -1),
-    (1207, 1207, -1),
-    (1209, 1209, -1),
-    (1211, 1211, -1),
-    (1213, 1213, -1),
-    (1215, 1215, -1),
-    (1218, 1218, -1),
-    (1220, 1220, -1),
-    (1222, 1222, -1),
-    (1224, 1224, -1),
-    (1226, 1226, -1),
-    (1228, 1228, -1),
-    (1230, 1230, -1),
-    (1231, 1231, -15),
-    (1233, 1233, -1),
-    (1235, 1235, -1),
-    (1237, 1237, -1),
-    (1239, 1239, -1),
-    (1241, 1241, -1),
-    (1243, 1243, -1),
-    (1245, 1245, -1),
-    (1247, 1247, -1),
-    (1249, 1249, -1),
-    (1251, 1251, -1),
-    (1253, 1253, -1),
-    (1255, 1255, -1),
-    (1257, 1257, -1),
-    (1259, 1259, -1),
-    (1261, 1261, -1),
-    (1263, 1263, -1),
-    (1265, 1265, -1),
-    (1267, 1267, -1),
-    (1269, 1269, -1),
-    (1271, 1271, -1),
-    (1273, 1273, -1),
-    (1275, 1275, -1),
-    (1277, 1277, -1),
-    (1279, 1279, -1),
-    (1281, 1281, -1),
-    (1283, 1283, -1),
-    (1285, 1285, -1),
-    (1287, 1287, -1),
-    (1289, 1289, -1),
-    (1291, 1291, -1),
-    (1293, 1293, -1),
-    (1295, 1295, -1),
-    (1297, 1297, -1),
-    (1299, 1299, -1),
-    (1301, 1301, -1),
-    (1303, 1303, -1),
-    (1305, 1305, -1),
-    (1307, 1307, -1),
-    (1309, 1309, -1),
-    (1311, 1311, -1),
-    (1313, 1313, -1),
-    (1315, 1315, -1),
-    (1317, 1317, -1),
-    (1319, 1319, -1),
-    (1321, 1321, -1),
-    (1323, 1323, -1),
-    (1325, 1325, -1),
-    (1327, 1327, -1),
-    (1377, 1414, -48),
-    (4304, 4346, 3008),
-    (4349, 4351, 3008),
-    (5112, 5117, -8),
-    (7296, 7296, -6254),
-    (7297, 7297, -6253),
-    (7298, 7298, -6244),
-    (7299, 7300, -6242),
-    (7301, 7301, -6243),
-    (7302, 7302, -6236),
-    (7303, 7303, -6181),
-    (7304, 7304, 35266),
-    (7306, 7306, -1),
-    (7545, 7545, 35332),
-    (7549, 7549, 3814),
-    (7566, 7566, 35384),
-    (7681, 7681, -1),
-    (7683, 7683, -1),
-    (7685, 7685, -1),
-    (7687, 7687, -1),
-    (7689, 7689, -1),
-    (7691, 7691, -1),
-    (7693, 7693, -1),
-    (7695, 7695, -1),
-    (7697, 7697, -1),
-    (7699, 7699, -1),
-    (7701, 7701, -1),
-    (7703, 7703, -1),
-    (7705, 7705, -1),
-    (7707, 7707, -1),
-    (7709, 7709, -1),
-    (7711, 7711, -1),
-    (7713, 7713, -1),
-    (7715, 7715, -1),
-    (7717, 7717, -1),
-    (7719, 7719, -1),
-    (7721, 7721, -1),
-    (7723, 7723, -1),
-    (7725, 7725, -1),
-    (7727, 7727, -1),
-    (7729, 7729, -1),
-    (7731, 7731, -1),
-    (7733, 7733, -1),
-    (7735, 7735, -1),
-    (7737, 7737, -1),
-    (7739, 7739, -1),
-    (7741, 7741, -1),
-    (7743, 7743, -1),
-    (7745, 7745, -1),
-    (7747, 7747, -1),
-    (7749, 7749, -1),
-    (7751, 7751, -1),
-    (7753, 7753, -1),
-    (7755, 7755, -1),
-    (7757, 7757, -1),
-    (7759, 7759, -1),
-    (7761, 7761, -1),
-    (7763, 7763, -1),
-    (7765, 7765, -1),
-    (7767, 7767, -1),
-    (7769, 7769, -1),
-    (7771, 7771, -1),
-    (7773, 7773, -1),
-    (7775, 7775, -1),
-    (7777, 7777, -1),
-    (7779, 7779, -1),
-    (7781, 7781, -1),
-    (7783, 7783, -1),
-    (7785, 7785, -1),
-    (7787, 7787, -1),
-    (7789, 7789, -1),
-    (7791, 7791, -1),
-    (7793, 7793, -1),
-    (7795, 7795, -1),
-    (7797, 7797, -1),
-    (7799, 7799, -1),
-    (7801, 7801, -1),
-    (7803, 7803, -1),
-    (7805, 7805, -1),
-    (7807, 7807, -1),
-    (7809, 7809, -1),
-    (7811, 7811, -1),
-    (7813, 7813, -1),
-    (7815, 7815, -1),
-    (7817, 7817, -1),
-    (7819, 7819, -1),
-    (7821, 7821, -1),
-    (7823, 7823, -1),
-    (7825, 7825, -1),
-    (7827, 7827, -1),
-    (7829, 7829, -1),
-    (7835, 7835, -59),
-    (7841, 7841, -1),
-    (7843, 7843, -1),
-    (7845, 7845, -1),
-    (7847, 7847, -1),
-    (7849, 7849, -1),
-    (7851, 7851, -1),
-    (7853, 7853, -1),
-    (7855, 7855, -1),
-    (7857, 7857, -1),
-    (7859, 7859, -1),
-    (7861, 7861, -1),
-    (7863, 7863, -1),
-    (7865, 7865, -1),
-    (7867, 7867, -1),
-    (7869, 7869, -1),
-    (7871, 7871, -1),
-    (7873, 7873, -1),
-    (7875, 7875, -1),
-    (7877, 7877, -1),
-    (7879, 7879, -1),
-    (7881, 7881, -1),
-    (7883, 7883, -1),
-    (7885, 7885, -1),
-    (7887, 7887, -1),
-    (7889, 7889, -1),
-    (7891, 7891, -1),
-    (7893, 7893, -1),
-    (7895, 7895, -1),
-    (7897, 7897, -1),
-    (7899, 7899, -1),
-    (7901, 7901, -1),
-    (7903, 7903, -1),
-    (7905, 7905, -1),
-    (7907, 7907, -1),
-    (7909, 7909, -1),
-    (7911, 7911, -1),
-    (7913, 7913, -1),
-    (7915, 7915, -1),
-    (7917, 7917, -1),
-    (7919, 7919, -1),
-    (7921, 7921, -1),
-    (7923, 7923, -1),
-    (7925, 7925, -1),
-    (7927, 7927, -1),
-    (7929, 7929, -1),
-    (7931, 7931, -1),
-    (7933, 7933, -1),
-    (7935, 7935, -1),
-    (7936, 7943, 8),
-    (7952, 7957, 8),
-    (7968, 7975, 8),
-    (7984, 7991, 8),
-    (8000, 8005, 8),
-    (8017, 8017, 8),
-    (8019, 8019, 8),
-    (8021, 8021, 8),
-    (8023, 8023, 8),
-    (8032, 8039, 8),
-    (8048, 8049, 74),
-    (8050, 8053, 86),
-    (8054, 8055, 100),
-    (8056, 8057, 128),
-    (8058, 8059, 112),
-    (8060, 8061, 126),
-    (8112, 8113, 8),
-    (8126, 8126, -7205),
-    (8144, 8145, 8),
-    (8160, 8161, 8),
-    (8165, 8165, 7),
-    (8526, 8526, -28),
-    (8560, 8575, -16),
-    (8580, 8580, -1),
-    (9424, 9449, -26),
-    (11312, 11359, -48),
-    (11361, 11361, -1),
-    (11365, 11365, -10795),
-    (11366, 11366, -10792),
-    (11368, 11368, -1),
-    (11370, 11370, -1),
-    (11372, 11372, -1),
-    (11379, 11379, -1),
-    (11382, 11382, -1),
-    (11393, 11393, -1),
-    (11395, 11395, -1),
-    (11397, 11397, -1),
-    (11399, 11399, -1),
-    (11401, 11401, -1),
-    (11403, 11403, -1),
-    (11405, 11405, -1),
-    (11407, 11407, -1),
-    (11409, 11409, -1),
-    (11411, 11411, -1),
-    (11413, 11413, -1),
-    (11415, 11415, -1),
-    (11417, 11417, -1),
-    (11419, 11419, -1),
-    (11421, 11421, -1),
-    (11423, 11423, -1),
-    (11425, 11425, -1),
-    (11427, 11427, -1),
-    (11429, 11429, -1),
-    (11431, 11431, -1),
-    (11433, 11433, -1),
-    (11435, 11435, -1),
-    (11437, 11437, -1),
-    (11439, 11439, -1),
-    (11441, 11441, -1),
-    (11443, 11443, -1),
-    (11445, 11445, -1),
-    (11447, 11447, -1),
-    (11449, 11449, -1),
-    (11451, 11451, -1),
-    (11453, 11453, -1),
-    (11455, 11455, -1),
-    (11457, 11457, -1),
-    (11459, 11459, -1),
-    (11461, 11461, -1),
-    (11463, 11463, -1),
-    (11465, 11465, -1),
-    (11467, 11467, -1),
-    (11469, 11469, -1),
-    (11471, 11471, -1),
-    (11473, 11473, -1),
-    (11475, 11475, -1),
-    (11477, 11477, -1),
-    (11479, 11479, -1),
-    (11481, 11481, -1),
-    (11483, 11483, -1),
-    (11485, 11485, -1),
-    (11487, 11487, -1),
-    (11489, 11489, -1),
-    (11491, 11491, -1),
-    (11500, 11500, -1),
-    (11502, 11502, -1),
-    (11507, 11507, -1),
-    (11520, 11557, -7264),
-    (11559, 11559, -7264),
-    (11565, 11565, -7264),
-    (42561, 42561, -1),
-    (42563, 42563, -1),
-    (42565, 42565, -1),
-    (42567, 42567, -1),
-    (42569, 42569, -1),
-    (42571, 42571, -1),
-    (42573, 42573, -1),
-    (42575, 42575, -1),
-    (42577, 42577, -1),
-    (42579, 42579, -1),
-    (42581, 42581, -1),
-    (42583, 42583, -1),
-    (42585, 42585, -1),
-    (42587, 42587, -1),
-    (42589, 42589, -1),
-    (42591, 42591, -1),
-    (42593, 42593, -1),
-    (42595, 42595, -1),
-    (42597, 42597, -1),
-    (42599, 42599, -1),
-    (42601, 42601, -1),
-    (42603, 42603, -1),
-    (42605, 42605, -1),
-    (42625, 42625, -1),
-    (42627, 42627, -1),
-    (42629, 42629, -1),
-    (42631, 42631, -1),
-    (42633, 42633, -1),
-    (42635, 42635, -1),
-    (42637, 42637, -1),
-    (42639, 42639, -1),
-    (42641, 42641, -1),
-    (42643, 42643, -1),
-    (42645, 42645, -1),
-    (42647, 42647, -1),
-    (42649, 42649, -1),
-    (42651, 42651, -1),
-    (42787, 42787, -1),
-    (42789, 42789, -1),
-    (42791, 42791, -1),
-    (42793, 42793, -1),
-    (42795, 42795, -1),
-    (42797, 42797, -1),
-    (42799, 42799, -1),
-    (42803, 42803, -1),
-    (42805, 42805, -1),
-    (42807, 42807, -1),
-    (42809, 42809, -1),
-    (42811, 42811, -1),
-    (42813, 42813, -1),
-    (42815, 42815, -1),
-    (42817, 42817, -1),
-    (42819, 42819, -1),
-    (42821, 42821, -1),
-    (42823, 42823, -1),
-    (42825, 42825, -1),
-    (42827, 42827, -1),
-    (42829, 42829, -1),
-    (42831, 42831, -1),
-    (42833, 42833, -1),
-    (42835, 42835, -1),
-    (42837, 42837, -1),
-    (42839, 42839, -1),
-    (42841, 42841, -1),
-    (42843, 42843, -1),
-    (42845, 42845, -1),
-    (42847, 42847, -1),
-    (42849, 42849, -1),
-    (42851, 42851, -1),
-    (42853, 42853, -1),
-    (42855, 42855, -1),
-    (42857, 42857, -1),
-    (42859, 42859, -1),
-    (42861, 42861, -1),
-    (42863, 42863, -1),
-    (42874, 42874, -1),
-    (42876, 42876, -1),
-    (42879, 42879, -1),
-    (42881, 42881, -1),
-    (42883, 42883, -1),
-    (42885, 42885, -1),
-    (42887, 42887, -1),
-    (42892, 42892, -1),
-    (42897, 42897, -1),
-    (42899, 42899, -1),
-    (42900, 42900, 48),
-    (42903, 42903, -1),
-    (42905, 42905, -1),
-    (42907, 42907, -1),
-    (42909, 42909, -1),
-    (42911, 42911, -1),
-    (42913, 42913, -1),
-    (42915, 42915, -1),
-    (42917, 42917, -1),
-    (42919, 42919, -1),
-    (42921, 42921, -1),
-    (42933, 42933, -1),
-    (42935, 42935, -1),
-    (42937, 42937, -1),
-    (42939, 42939, -1),
-    (42941, 42941, -1),
-    (42943, 42943, -1),
-    (42945, 42945, -1),
-    (42947, 42947, -1),
-    (42952, 42952, -1),
-    (42954, 42954, -1),
-    (42957, 42957, -1),
-    (42959, 42959, -1),
-    (42961, 42961, -1),
-    (42963, 42963, -1),
-    (42965, 42965, -1),
-    (42967, 42967, -1),
-    (42969, 42969, -1),
-    (42971, 42971, -1),
-    (42998, 42998, -1),
-    (43859, 43859, -928),
-    (43888, 43967, -38864),
-    (65345, 65370, -32),
-    (66600, 66639, -40),
-    (66776, 66811, -40),
-    (66967, 66977, -39),
-    (66979, 66993, -39),
-    (66995, 67001, -39),
-    (67003, 67004, -39),
-    (68800, 68850, -64),
-    (68976, 68997, -32),
-    (71872, 71903, -32),
-    (93792, 93823, -32),
-    (93883, 93907, -27),
-    (125218, 125251, -34),
-];
-pub(in crate::codegen::wasm_gc) const LOWER_EXPAND: &[Expansion] = &[(304, [105, 775, 0])];
-pub(in crate::codegen::wasm_gc) const UPPER_EXPAND: &[Expansion] = &[
-    (223, [83, 83, 0]),
-    (329, [700, 78, 0]),
-    (496, [74, 780, 0]),
-    (912, [921, 776, 769]),
-    (944, [933, 776, 769]),
-    (1415, [1333, 1362, 0]),
-    (7830, [72, 817, 0]),
-    (7831, [84, 776, 0]),
-    (7832, [87, 778, 0]),
-    (7833, [89, 778, 0]),
-    (7834, [65, 702, 0]),
-    (8016, [933, 787, 0]),
-    (8018, [933, 787, 768]),
-    (8020, [933, 787, 769]),
-    (8022, [933, 787, 834]),
-    (8064, [7944, 921, 0]),
-    (8065, [7945, 921, 0]),
-    (8066, [7946, 921, 0]),
-    (8067, [7947, 921, 0]),
-    (8068, [7948, 921, 0]),
-    (8069, [7949, 921, 0]),
-    (8070, [7950, 921, 0]),
-    (8071, [7951, 921, 0]),
-    (8072, [7944, 921, 0]),
-    (8073, [7945, 921, 0]),
-    (8074, [7946, 921, 0]),
-    (8075, [7947, 921, 0]),
-    (8076, [7948, 921, 0]),
-    (8077, [7949, 921, 0]),
-    (8078, [7950, 921, 0]),
-    (8079, [7951, 921, 0]),
-    (8080, [7976, 921, 0]),
-    (8081, [7977, 921, 0]),
-    (8082, [7978, 921, 0]),
-    (8083, [7979, 921, 0]),
-    (8084, [7980, 921, 0]),
-    (8085, [7981, 921, 0]),
-    (8086, [7982, 921, 0]),
-    (8087, [7983, 921, 0]),
-    (8088, [7976, 921, 0]),
-    (8089, [7977, 921, 0]),
-    (8090, [7978, 921, 0]),
-    (8091, [7979, 921, 0]),
-    (8092, [7980, 921, 0]),
-    (8093, [7981, 921, 0]),
-    (8094, [7982, 921, 0]),
-    (8095, [7983, 921, 0]),
-    (8096, [8040, 921, 0]),
-    (8097, [8041, 921, 0]),
-    (8098, [8042, 921, 0]),
-    (8099, [8043, 921, 0]),
-    (8100, [8044, 921, 0]),
-    (8101, [8045, 921, 0]),
-    (8102, [8046, 921, 0]),
-    (8103, [8047, 921, 0]),
-    (8104, [8040, 921, 0]),
-    (8105, [8041, 921, 0]),
-    (8106, [8042, 921, 0]),
-    (8107, [8043, 921, 0]),
-    (8108, [8044, 921, 0]),
-    (8109, [8045, 921, 0]),
-    (8110, [8046, 921, 0]),
-    (8111, [8047, 921, 0]),
-    (8114, [8122, 921, 0]),
-    (8115, [913, 921, 0]),
-    (8116, [902, 921, 0]),
-    (8118, [913, 834, 0]),
-    (8119, [913, 834, 921]),
-    (8124, [913, 921, 0]),
-    (8130, [8138, 921, 0]),
-    (8131, [919, 921, 0]),
-    (8132, [905, 921, 0]),
-    (8134, [919, 834, 0]),
-    (8135, [919, 834, 921]),
-    (8140, [919, 921, 0]),
-    (8146, [921, 776, 768]),
-    (8147, [921, 776, 769]),
-    (8150, [921, 834, 0]),
-    (8151, [921, 776, 834]),
-    (8162, [933, 776, 768]),
-    (8163, [933, 776, 769]),
-    (8164, [929, 787, 0]),
-    (8166, [933, 834, 0]),
-    (8167, [933, 776, 834]),
-    (8178, [8186, 921, 0]),
-    (8179, [937, 921, 0]),
-    (8180, [911, 921, 0]),
-    (8182, [937, 834, 0]),
-    (8183, [937, 834, 921]),
-    (8188, [937, 921, 0]),
-    (64256, [70, 70, 0]),
-    (64257, [70, 73, 0]),
-    (64258, [70, 76, 0]),
-    (64259, [70, 70, 73]),
-    (64260, [70, 70, 76]),
-    (64261, [83, 84, 0]),
-    (64262, [83, 84, 0]),
-    (64275, [1348, 1350, 0]),
-    (64276, [1348, 1333, 0]),
-    (64277, [1348, 1339, 0]),
-    (64278, [1358, 1350, 0]),
-    (64279, [1348, 1341, 0]),
-];
-pub(in crate::codegen::wasm_gc) const CASED: &[Range] = &[
-    (65, 90),
-    (97, 122),
-    (170, 170),
-    (181, 181),
-    (186, 186),
-    (192, 214),
-    (216, 246),
-    (248, 442),
-    (444, 447),
-    (452, 659),
-    (662, 687),
-    (880, 883),
-    (886, 887),
-    (891, 893),
-    (895, 895),
-    (902, 902),
-    (904, 906),
-    (908, 908),
-    (910, 929),
-    (931, 1013),
-    (1015, 1153),
-    (1162, 1327),
-    (1329, 1366),
-    (1376, 1416),
-    (4256, 4293),
-    (4295, 4295),
-    (4301, 4301),
-    (4304, 4346),
-    (4349, 4351),
-    (5024, 5109),
-    (5112, 5117),
-    (7296, 7306),
-    (7312, 7354),
-    (7357, 7359),
-    (7424, 7467),
-    (7531, 7543),
-    (7545, 7578),
-    (7680, 7957),
-    (7960, 7965),
-    (7968, 8005),
-    (8008, 8013),
-    (8016, 8023),
-    (8025, 8025),
-    (8027, 8027),
-    (8029, 8029),
-    (8031, 8061),
-    (8064, 8116),
-    (8118, 8124),
-    (8126, 8126),
-    (8130, 8132),
-    (8134, 8140),
-    (8144, 8147),
-    (8150, 8155),
-    (8160, 8172),
-    (8178, 8180),
-    (8182, 8188),
-    (8450, 8450),
-    (8455, 8455),
-    (8458, 8467),
-    (8469, 8469),
-    (8473, 8477),
-    (8484, 8484),
-    (8486, 8486),
-    (8488, 8488),
-    (8490, 8493),
-    (8495, 8500),
-    (8505, 8505),
-    (8508, 8511),
-    (8517, 8521),
-    (8526, 8526),
-    (8544, 8575),
-    (8579, 8580),
-    (9398, 9449),
-    (11264, 11387),
-    (11390, 11492),
-    (11499, 11502),
-    (11506, 11507),
-    (11520, 11557),
-    (11559, 11559),
-    (11565, 11565),
-    (42560, 42605),
-    (42624, 42651),
-    (42786, 42863),
-    (42865, 42887),
-    (42891, 42894),
-    (42896, 42972),
-    (42997, 42998),
-    (43002, 43002),
-    (43824, 43866),
-    (43872, 43880),
-    (43888, 43967),
-    (64256, 64262),
-    (64275, 64279),
-    (65313, 65338),
-    (65345, 65370),
-    (66560, 66639),
-    (66736, 66771),
-    (66776, 66811),
-    (66928, 66938),
-    (66940, 66954),
-    (66956, 66962),
-    (66964, 66965),
-    (66967, 66977),
-    (66979, 66993),
-    (66995, 67001),
-    (67003, 67004),
-    (68736, 68786),
-    (68800, 68850),
-    (68944, 68965),
-    (68976, 68997),
-    (71840, 71903),
-    (93760, 93823),
-    (93856, 93880),
-    (93883, 93907),
-    (119808, 119892),
-    (119894, 119964),
-    (119966, 119967),
-    (119970, 119970),
-    (119973, 119974),
-    (119977, 119980),
-    (119982, 119993),
-    (119995, 119995),
-    (119997, 120003),
-    (120005, 120069),
-    (120071, 120074),
-    (120077, 120084),
-    (120086, 120092),
-    (120094, 120121),
-    (120123, 120126),
-    (120128, 120132),
-    (120134, 120134),
-    (120138, 120144),
-    (120146, 120485),
-    (120488, 120512),
-    (120514, 120538),
-    (120540, 120570),
-    (120572, 120596),
-    (120598, 120628),
-    (120630, 120654),
-    (120656, 120686),
-    (120688, 120712),
-    (120714, 120744),
-    (120746, 120770),
-    (120772, 120779),
-    (122624, 122633),
-    (122635, 122654),
-    (122661, 122666),
-    (125184, 125251),
-    (127280, 127305),
-    (127312, 127337),
-    (127344, 127369),
-];
-pub(in crate::codegen::wasm_gc) const CASE_IGNORABLE: &[Range] = &[
-    (39, 39),
-    (46, 46),
-    (58, 58),
-    (94, 94),
-    (96, 96),
-    (168, 168),
-    (173, 173),
-    (175, 175),
-    (180, 180),
-    (183, 184),
-    (688, 879),
-    (884, 885),
-    (890, 890),
-    (900, 901),
-    (903, 903),
-    (1155, 1161),
-    (1369, 1369),
-    (1375, 1375),
-    (1425, 1469),
-    (1471, 1471),
-    (1473, 1474),
-    (1476, 1477),
-    (1479, 1479),
-    (1524, 1524),
-    (1536, 1541),
-    (1552, 1562),
-    (1564, 1564),
-    (1600, 1600),
-    (1611, 1631),
-    (1648, 1648),
-    (1750, 1757),
-    (1759, 1768),
-    (1770, 1773),
-    (1807, 1807),
-    (1809, 1809),
-    (1840, 1866),
-    (1958, 1968),
-    (2027, 2037),
-    (2042, 2042),
-    (2045, 2045),
-    (2070, 2093),
-    (2137, 2139),
-    (2184, 2184),
-    (2192, 2193),
-    (2199, 2207),
-    (2249, 2306),
-    (2362, 2362),
-    (2364, 2364),
-    (2369, 2376),
-    (2381, 2381),
-    (2385, 2391),
-    (2402, 2403),
-    (2417, 2417),
-    (2433, 2433),
-    (2492, 2492),
-    (2497, 2500),
-    (2509, 2509),
-    (2530, 2531),
-    (2558, 2558),
-    (2561, 2562),
-    (2620, 2620),
-    (2625, 2626),
-    (2631, 2632),
-    (2635, 2637),
-    (2641, 2641),
-    (2672, 2673),
-    (2677, 2677),
-    (2689, 2690),
-    (2748, 2748),
-    (2753, 2757),
-    (2759, 2760),
-    (2765, 2765),
-    (2786, 2787),
-    (2810, 2815),
-    (2817, 2817),
-    (2876, 2876),
-    (2879, 2879),
-    (2881, 2884),
-    (2893, 2893),
-    (2901, 2902),
-    (2914, 2915),
-    (2946, 2946),
-    (3008, 3008),
-    (3021, 3021),
-    (3072, 3072),
-    (3076, 3076),
-    (3132, 3132),
-    (3134, 3136),
-    (3142, 3144),
-    (3146, 3149),
-    (3157, 3158),
-    (3170, 3171),
-    (3201, 3201),
-    (3260, 3260),
-    (3263, 3263),
-    (3270, 3270),
-    (3276, 3277),
-    (3298, 3299),
-    (3328, 3329),
-    (3387, 3388),
-    (3393, 3396),
-    (3405, 3405),
-    (3426, 3427),
-    (3457, 3457),
-    (3530, 3530),
-    (3538, 3540),
-    (3542, 3542),
-    (3633, 3633),
-    (3636, 3642),
-    (3654, 3662),
-    (3761, 3761),
-    (3764, 3772),
-    (3782, 3782),
-    (3784, 3790),
-    (3864, 3865),
-    (3893, 3893),
-    (3895, 3895),
-    (3897, 3897),
-    (3953, 3966),
-    (3968, 3972),
-    (3974, 3975),
-    (3981, 3991),
-    (3993, 4028),
-    (4038, 4038),
-    (4141, 4144),
-    (4146, 4151),
-    (4153, 4154),
-    (4157, 4158),
-    (4184, 4185),
-    (4190, 4192),
-    (4209, 4212),
-    (4226, 4226),
-    (4229, 4230),
-    (4237, 4237),
-    (4253, 4253),
-    (4348, 4348),
-    (4957, 4959),
-    (5906, 5908),
-    (5938, 5939),
-    (5970, 5971),
-    (6002, 6003),
-    (6068, 6069),
-    (6071, 6077),
-    (6086, 6086),
-    (6089, 6099),
-    (6103, 6103),
-    (6109, 6109),
-    (6155, 6159),
-    (6211, 6211),
-    (6277, 6278),
-    (6313, 6313),
-    (6432, 6434),
-    (6439, 6440),
-    (6450, 6450),
-    (6457, 6459),
-    (6679, 6680),
-    (6683, 6683),
-    (6742, 6742),
-    (6744, 6750),
-    (6752, 6752),
-    (6754, 6754),
-    (6757, 6764),
-    (6771, 6780),
-    (6783, 6783),
-    (6823, 6823),
-    (6832, 6877),
-    (6880, 6891),
-    (6912, 6915),
-    (6964, 6964),
-    (6966, 6970),
-    (6972, 6972),
-    (6978, 6978),
-    (7019, 7027),
-    (7040, 7041),
-    (7074, 7077),
-    (7080, 7081),
-    (7083, 7085),
-    (7142, 7142),
-    (7144, 7145),
-    (7149, 7149),
-    (7151, 7153),
-    (7212, 7219),
-    (7222, 7223),
-    (7288, 7293),
-    (7376, 7378),
-    (7380, 7392),
-    (7394, 7400),
-    (7405, 7405),
-    (7412, 7412),
-    (7416, 7417),
-    (7468, 7530),
-    (7544, 7544),
-    (7579, 7679),
-    (8125, 8125),
-    (8127, 8129),
-    (8141, 8143),
-    (8157, 8159),
-    (8173, 8175),
-    (8189, 8190),
-    (8203, 8207),
-    (8216, 8217),
-    (8228, 8228),
-    (8231, 8231),
-    (8234, 8238),
-    (8288, 8292),
-    (8294, 8303),
-    (8305, 8305),
-    (8319, 8319),
-    (8336, 8348),
-    (8400, 8432),
-    (11388, 11389),
-    (11503, 11505),
-    (11631, 11631),
-    (11647, 11647),
-    (11744, 11775),
-    (11823, 11823),
-    (12293, 12293),
-    (12330, 12333),
-    (12337, 12341),
-    (12347, 12347),
-    (12441, 12446),
-    (12540, 12542),
-    (40981, 40981),
-    (42232, 42237),
-    (42508, 42508),
-    (42607, 42610),
-    (42612, 42621),
-    (42623, 42623),
-    (42652, 42655),
-    (42736, 42737),
-    (42752, 42785),
-    (42864, 42864),
-    (42888, 42890),
-    (42993, 42996),
-    (43000, 43001),
-    (43010, 43010),
-    (43014, 43014),
-    (43019, 43019),
-    (43045, 43046),
-    (43052, 43052),
-    (43204, 43205),
-    (43232, 43249),
-    (43263, 43263),
-    (43302, 43309),
-    (43335, 43345),
-    (43392, 43394),
-    (43443, 43443),
-    (43446, 43449),
-    (43452, 43453),
-    (43471, 43471),
-    (43493, 43494),
-    (43561, 43566),
-    (43569, 43570),
-    (43573, 43574),
-    (43587, 43587),
-    (43596, 43596),
-    (43632, 43632),
-    (43644, 43644),
-    (43696, 43696),
-    (43698, 43700),
-    (43703, 43704),
-    (43710, 43711),
-    (43713, 43713),
-    (43741, 43741),
-    (43756, 43757),
-    (43763, 43764),
-    (43766, 43766),
-    (43867, 43871),
-    (43881, 43883),
-    (44005, 44005),
-    (44008, 44008),
-    (44013, 44013),
-    (64286, 64286),
-    (64434, 64450),
-    (65024, 65039),
-    (65043, 65043),
-    (65056, 65071),
-    (65106, 65106),
-    (65109, 65109),
-    (65279, 65279),
-    (65287, 65287),
-    (65294, 65294),
-    (65306, 65306),
-    (65342, 65342),
-    (65344, 65344),
-    (65392, 65392),
-    (65438, 65439),
-    (65507, 65507),
-    (65529, 65531),
-    (66045, 66045),
-    (66272, 66272),
-    (66422, 66426),
-    (67456, 67461),
-    (67463, 67504),
-    (67506, 67514),
-    (68097, 68099),
-    (68101, 68102),
-    (68108, 68111),
-    (68152, 68154),
-    (68159, 68159),
-    (68325, 68326),
-    (68900, 68903),
-    (68942, 68942),
-    (68969, 68973),
-    (68975, 68975),
-    (69291, 69292),
-    (69317, 69317),
-    (69370, 69375),
-    (69446, 69456),
-    (69506, 69509),
-    (69633, 69633),
-    (69688, 69702),
-    (69744, 69744),
-    (69747, 69748),
-    (69759, 69761),
-    (69811, 69814),
-    (69817, 69818),
-    (69821, 69821),
-    (69826, 69826),
-    (69837, 69837),
-    (69888, 69890),
-    (69927, 69931),
-    (69933, 69940),
-    (70003, 70003),
-    (70016, 70017),
-    (70070, 70078),
-    (70089, 70092),
-    (70095, 70095),
-    (70191, 70193),
-    (70196, 70196),
-    (70198, 70199),
-    (70206, 70206),
-    (70209, 70209),
-    (70367, 70367),
-    (70371, 70378),
-    (70400, 70401),
-    (70459, 70460),
-    (70464, 70464),
-    (70502, 70508),
-    (70512, 70516),
-    (70587, 70592),
-    (70606, 70606),
-    (70608, 70608),
-    (70610, 70610),
-    (70625, 70626),
-    (70712, 70719),
-    (70722, 70724),
-    (70726, 70726),
-    (70750, 70750),
-    (70835, 70840),
-    (70842, 70842),
-    (70847, 70848),
-    (70850, 70851),
-    (71090, 71093),
-    (71100, 71101),
-    (71103, 71104),
-    (71132, 71133),
-    (71219, 71226),
-    (71229, 71229),
-    (71231, 71232),
-    (71339, 71339),
-    (71341, 71341),
-    (71344, 71349),
-    (71351, 71351),
-    (71453, 71453),
-    (71455, 71455),
-    (71458, 71461),
-    (71463, 71467),
-    (71727, 71735),
-    (71737, 71738),
-    (71995, 71996),
-    (71998, 71998),
-    (72003, 72003),
-    (72148, 72151),
-    (72154, 72155),
-    (72160, 72160),
-    (72193, 72202),
-    (72243, 72248),
-    (72251, 72254),
-    (72263, 72263),
-    (72273, 72278),
-    (72281, 72283),
-    (72330, 72342),
-    (72344, 72345),
-    (72544, 72544),
-    (72546, 72548),
-    (72550, 72550),
-    (72752, 72758),
-    (72760, 72765),
-    (72767, 72767),
-    (72850, 72871),
-    (72874, 72880),
-    (72882, 72883),
-    (72885, 72886),
-    (73009, 73014),
-    (73018, 73018),
-    (73020, 73021),
-    (73023, 73029),
-    (73031, 73031),
-    (73104, 73105),
-    (73109, 73109),
-    (73111, 73111),
-    (73177, 73177),
-    (73459, 73460),
-    (73472, 73473),
-    (73526, 73530),
-    (73536, 73536),
-    (73538, 73538),
-    (73562, 73562),
-    (78896, 78912),
-    (78919, 78933),
-    (90398, 90409),
-    (90413, 90415),
-    (92912, 92916),
-    (92976, 92982),
-    (92992, 92995),
-    (93504, 93506),
-    (93547, 93548),
-    (94031, 94031),
-    (94095, 94111),
-    (94176, 94177),
-    (94179, 94180),
-    (94194, 94195),
-    (110576, 110579),
-    (110581, 110587),
-    (110589, 110590),
-    (113821, 113822),
-    (113824, 113827),
-    (118528, 118573),
-    (118576, 118598),
-    (119143, 119145),
-    (119155, 119170),
-    (119173, 119179),
-    (119210, 119213),
-    (119362, 119364),
-    (121344, 121398),
-    (121403, 121452),
-    (121461, 121461),
-    (121476, 121476),
-    (121499, 121503),
-    (121505, 121519),
-    (122880, 122886),
-    (122888, 122904),
-    (122907, 122913),
-    (122915, 122916),
-    (122918, 122922),
-    (122928, 122989),
-    (123023, 123023),
-    (123184, 123197),
-    (123566, 123566),
-    (123628, 123631),
-    (124139, 124143),
-    (124398, 124399),
-    (124643, 124643),
-    (124646, 124646),
-    (124654, 124655),
-    (124661, 124661),
-    (124671, 124671),
-    (125136, 125142),
-    (125252, 125259),
-    (127995, 127999),
-    (917505, 917505),
-    (917536, 917631),
-    (917760, 917999),
-];
-// @end generated
+use crate::codegen::unicode_case::{
+    CASE_IGNORABLE, CASED, Expansion, LOWER_EXPAND, LOWER_SIMPLE, Range, SimpleRun, UPPER_EXPAND,
+    UPPER_SIMPLE,
+};
+pub(in crate::codegen::wasm_gc) use crate::codegen::unicode_case::{
+    MAX_LOWER_GROWTH, MAX_UPPER_GROWTH,
+};
 
 /// Bytes per record in the emitted blob. Every field is a
 /// little-endian `u24`, so one read routine in WAT serves all four
@@ -2277,197 +131,6 @@ pub(in crate::codegen::wasm_gc) fn build_blob(need_lower: bool, need_upper: bool
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[derive(Default)]
-    struct Derived {
-        lower_simple: Vec<SimpleRun>,
-        upper_simple: Vec<SimpleRun>,
-        lower_expand: Vec<Expansion>,
-        upper_expand: Vec<Expansion>,
-        cased: Vec<Range>,
-        case_ignorable: Vec<Range>,
-        max_lower_growth: u32,
-        max_upper_growth: u32,
-    }
-
-    /// `Σ` and `ς` — the two scalars the context probes are built from.
-    const SIGMA: char = '\u{03A3}';
-    const FINAL_SIGMA: char = '\u{03C2}';
-
-    fn cased(c: char) -> bool {
-        format!("{c}{SIGMA}").to_lowercase().ends_with(FINAL_SIGMA)
-    }
-
-    fn case_ignorable(c: char) -> bool {
-        !cased(c) && format!("a{c}{SIGMA}").to_lowercase().ends_with(FINAL_SIGMA)
-    }
-
-    fn push_run(runs: &mut Vec<SimpleRun>, cp: u32, delta: i32) {
-        match runs.last_mut() {
-            Some(last) if last.1 + 1 == cp && last.2 == delta => last.1 = cp,
-            _ => runs.push((cp, cp, delta)),
-        }
-    }
-
-    fn push_range(ranges: &mut Vec<Range>, cp: u32) {
-        match ranges.last_mut() {
-            Some(last) if last.1 + 1 == cp => last.1 = cp,
-            _ => ranges.push((cp, cp)),
-        }
-    }
-
-    fn expansion(mapped: &[char]) -> [u32; 3] {
-        let mut out = [0u32; 3];
-        for (slot, c) in out.iter_mut().zip(mapped) {
-            *slot = *c as u32;
-        }
-        out
-    }
-
-    fn growth(mapped: &[char], input: char) -> u32 {
-        let out: usize = mapped.iter().map(|c| c.len_utf8()).sum();
-        out.div_ceil(input.len_utf8()) as u32
-    }
-
-    /// Re-derive every table from `std` by walking the whole scalar
-    /// space. Surrogates have no `char` and are skipped, which also
-    /// stops any run or range from spanning the surrogate block.
-    fn derive_from_std() -> Derived {
-        let mut d = Derived {
-            max_lower_growth: 1,
-            max_upper_growth: 1,
-            ..Derived::default()
-        };
-        for cp in 0u32..=0x10_FFFF {
-            let Some(c) = char::from_u32(cp) else {
-                continue;
-            };
-            let lower: Vec<char> = c.to_lowercase().collect();
-            let upper: Vec<char> = c.to_uppercase().collect();
-            d.max_lower_growth = d.max_lower_growth.max(growth(&lower, c));
-            d.max_upper_growth = d.max_upper_growth.max(growth(&upper, c));
-            match lower.as_slice() {
-                [one] if *one != c => push_run(&mut d.lower_simple, cp, *one as i32 - cp as i32),
-                [_] => {}
-                many => d.lower_expand.push((cp, expansion(many))),
-            }
-            match upper.as_slice() {
-                [one] if *one != c => push_run(&mut d.upper_simple, cp, *one as i32 - cp as i32),
-                [_] => {}
-                many => d.upper_expand.push((cp, expansion(many))),
-            }
-            if cased(c) {
-                push_range(&mut d.cased, cp);
-            } else if case_ignorable(c) {
-                push_range(&mut d.case_ignorable, cp);
-            }
-        }
-        d
-    }
-
-    fn sorted_and_disjoint(ranges: &[Range]) -> bool {
-        ranges
-            .windows(2)
-            .all(|w| w[0].0 <= w[0].1 && w[0].1 + 1 < w[1].0)
-    }
-
-    /// The drift guard. Red the day the toolchain's Unicode version
-    /// moves — regenerate with `dump_case_tables` and re-review.
-    #[test]
-    fn case_tables_match_std() {
-        let d = derive_from_std();
-        assert_eq!(d.lower_simple, LOWER_SIMPLE, "LOWER_SIMPLE");
-        assert_eq!(d.upper_simple, UPPER_SIMPLE, "UPPER_SIMPLE");
-        assert_eq!(d.lower_expand, LOWER_EXPAND, "LOWER_EXPAND");
-        assert_eq!(d.upper_expand, UPPER_EXPAND, "UPPER_EXPAND");
-        assert_eq!(d.cased, CASED, "CASED");
-        assert_eq!(d.case_ignorable, CASE_IGNORABLE, "CASE_IGNORABLE");
-        assert_eq!(d.max_lower_growth, MAX_LOWER_GROWTH, "MAX_LOWER_GROWTH");
-        assert_eq!(d.max_upper_growth, MAX_UPPER_GROWTH, "MAX_UPPER_GROWTH");
-        assert!(sorted_and_disjoint(CASED), "CASED must be sorted, disjoint");
-        assert!(
-            sorted_and_disjoint(CASE_IGNORABLE),
-            "CASE_IGNORABLE must be sorted, disjoint"
-        );
-        // The probes make the two sets disjoint; the helper relies on
-        // it (it tests ignorable first and only then cased).
-        for (first, last) in CASED {
-            assert!(
-                !CASE_IGNORABLE.iter().any(|(f, l)| first <= l && f <= last),
-                "CASED and CASE_IGNORABLE overlap around U+{first:04X}"
-            );
-        }
-        // Every table is binary-searched by its first field.
-        assert!(
-            LOWER_SIMPLE.windows(2).all(|w| w[0].1 < w[1].0),
-            "LOWER_SIMPLE must be sorted, disjoint"
-        );
-        assert!(
-            UPPER_SIMPLE.windows(2).all(|w| w[0].1 < w[1].0),
-            "UPPER_SIMPLE must be sorted, disjoint"
-        );
-        assert!(
-            LOWER_EXPAND.windows(2).all(|w| w[0].0 < w[1].0),
-            "LOWER_EXPAND must be sorted"
-        );
-        assert!(
-            UPPER_EXPAND.windows(2).all(|w| w[0].0 < w[1].0),
-            "UPPER_EXPAND must be sorted"
-        );
-        // `to2 == 0` is the "only two scalars" marker, so a genuine
-        // second scalar must never be U+0000.
-        assert!(
-            LOWER_EXPAND
-                .iter()
-                .chain(UPPER_EXPAND)
-                .all(|(_, to)| to[0] != 0 && to[1] != 0),
-            "expansion slots 0 and 1 are always present"
-        );
-        // 24-bit fields, sign-extended in WAT for the deltas.
-        assert!(
-            LOWER_SIMPLE
-                .iter()
-                .chain(UPPER_SIMPLE)
-                .all(|(_, _, delta)| (-0x80_0000..0x80_0000).contains(delta)),
-            "deltas must fit a signed 24-bit field"
-        );
-    }
-
-    /// Print the constants as Rust source. Ignored: it is the
-    /// regeneration path, not a check.
-    #[test]
-    #[ignore]
-    fn dump_case_tables() {
-        let d = derive_from_std();
-        println!("// @generated by `dump_case_tables` — see the module docs.");
-        println!(
-            "pub(in crate::codegen::wasm_gc) const LOWER_SIMPLE: &[SimpleRun] = &{:?};",
-            d.lower_simple
-        );
-        println!(
-            "pub(in crate::codegen::wasm_gc) const UPPER_SIMPLE: &[SimpleRun] = &{:?};",
-            d.upper_simple
-        );
-        println!(
-            "pub(in crate::codegen::wasm_gc) const LOWER_EXPAND: &[Expansion] = &{:?};",
-            d.lower_expand
-        );
-        println!(
-            "pub(in crate::codegen::wasm_gc) const UPPER_EXPAND: &[Expansion] = &{:?};",
-            d.upper_expand
-        );
-        println!(
-            "pub(in crate::codegen::wasm_gc) const CASED: &[Range] = &{:?};",
-            d.cased
-        );
-        println!(
-            "pub(in crate::codegen::wasm_gc) const CASE_IGNORABLE: &[Range] = &{:?};",
-            d.case_ignorable
-        );
-        println!("// MAX_LOWER_GROWTH = {}", d.max_lower_growth);
-        println!("// MAX_UPPER_GROWTH = {}", d.max_upper_growth);
-    }
-
     /// The blob layout the helper WAT is generated against.
     #[test]
     fn blob_spans_cover_the_bytes() {
